% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{betaHat}
\alias{betaHat}
\title{definition of beta hat. Not used in iterative process}
\usage{
betaHat(X, W, z)
}
\arguments{
\item{X}{matrix of coefficients}

\item{W}{diagonal... [EXPLAIN]}

\item{z}{vector of... [EXPLAIN]}
}
\description{
definition of beta hat. Not used in iterative process
}

