% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{betaInit}
\alias{betaInit}
\title{Initial guess of beta vector}
\usage{
betaInit(X, y)
}
\arguments{
\item{X}{matrix of covariates}

\item{y}{vector of outcomes}
}
\description{
Initial guess of beta vector
}

