% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{getCoef}
\alias{getCoef}
\alias{getCoef,GLMP-method}
\title{Set the generic for accessor (getter) for coefficients}
\usage{
getCoef(object)

\S4method{getCoef}{GLMP}(object)
}
\arguments{
\item{object}{object of class GLMP}
}
\description{
Set the generic for accessor (getter) for coefficients
}
\section{Methods (by class)}{
\itemize{
\item \code{GLMP}: 
}}

