% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-definition.R
\docType{methods}
\name{initialize,GLMP-method}
\alias{initialize,GLMP-method}
\title{Initialiser for GLMP objects}
\usage{
\S4method{initialize}{GLMP}(.Object, model, intercept, outcome, covariates,
  modelCoef)
}
\arguments{
\item{.Object}{object of class GLMP}

\item{model}{character string representing model}

\item{intercept}{logical. is intercept included?}

\item{outcome}{numerical vector of the outcome}

\item{covariates}{data frame holding numerical covariates}

\item{modelCoef}{matrix of the model coefficients}
}
\description{
Initialiser for GLMP objects
}

