% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{LL}
\alias{LL}
\title{liklihood function for binomial}
\usage{
LL(beta, X, y)
}
\arguments{
\item{beta}{vector of coefficients}

\item{X}{matrix of covariates}

\item{y}{dichotomous vector of outcomes (0, 1)}
}
\description{
liklihood function for binomial
}

