% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{show,GLMP-method}
\alias{show,GLMP-method}
\title{Show generic}
\usage{
\S4method{show}{GLMP}(object)
}
\arguments{
\item{object}{object of class GLMP}
}
\description{
Show generic
}

