% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{varBeta}
\alias{varBeta}
\title{Variance of coefficients}
\usage{
varBeta(X, W)
}
\arguments{
\item{X}{matrix of coefficients}

\item{W}{diagonal matrix of... [EXPLAIN]}
}
\description{
Variance of coefficients
}

