% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{zval}
\alias{zval}
\title{z vector}
\usage{
zval(beta, X, y)
}
\arguments{
\item{beta}{vector of coefficients}

\item{X}{matrix of covariates}

\item{y}{vecto of outcomes}
}
\description{
z vector
}

